MyProctor.ai-AI-BASED-SMART-ONLINE-EXAMINATION-PROCTORING-SYSYTEM/requirements.txt
Flask_WTF==1.2.1
pandas==2.1.4
six==1.16.0
tensorflow==2.15.0
WTForms_Components==0.10.5
stripe==2.27.0
deepface==0.0.87
Werkzeug==3.0.1
opencv_contrib_python==4.9.0.80
WTForms==3.1.2
Flask_MySQLdb==1.0.1
Flask==2.3.3
Flask_Mail==0.9.1
matplotlib==3.8.2
Flask_Cors==4.0.0
numpy==1.26.3
wget==3.2
python_docx==0.8.11
nltk==3.8.1
coolname==2.1.0
Pillow==10.1.0
